firmware flash

+++ App/AppSDT/FirmwareFlashDialog/firmwareflashdialog.cpp
@@ -4,12 +4,15 @@
 #include "sevdevice.h"
 #include "gtutils.h"
 #include "qttreemanager.h"
+#include "dbmanager.h"
+#include "deviceidhelper.h"
 
 #include <QFile>
 #include <QFileDialog>
 #include <QFileInfo>
 #include <QMessageBox>
 #include <QProgressBar>
+#include <QTextStream>
 #include <QDebug>
 
 #define FILENAME_XML_FLASHPRM "FlashPrm_AllAxis.xml"
@@ -24,16 +27,16 @@ FirmwareFlashDialog::FirmwareFlashDialog(QList<SevDevice *> &devList, QWidget *p
     ui(new Ui::FirmwareFlashDialog)
 {
     ui->setupUi(this);
+    m_devList = devList;
     uiInit();
-    createConnections();
-    m_devList = &devList;
     m_filePath = ".";
     m_desPath = "./DecompressdFiles";
-    QDir dir(m_desPath);
-    if (dir.exists()) {
-        deleteDir(m_desPath);
-    }
-    dir.mkdir(m_desPath);
+//    QDir dir(m_desPath);
+//    if (dir.exists()) {
+//        deleteDir(m_desPath);
+//    }
+//    dir.mkdir(m_desPath);
+    createConnections();
 }
 
 FirmwareFlashDialog::~FirmwareFlashDialog()
@@ -45,15 +48,18 @@ void FirmwareFlashDialog::uiInit()
 {
     ui->progressBar_firm->setVisible(false);
     ui->widget_firm->setVisible(false);
-    for (int i = 0; i < m_devList->count(); i++) {
-        ui->comboBox_firm->addItem(m_devList->at(i)->modelName());
+    for (int i = 0; i < m_devList.length(); i++) {
+        ui->comboBox_firm->addItem(m_devList.at(i)->modelName());
     }
+    ui->comboBox_firm->setCurrentIndex(0);
+    //m_devList
 }
 
 void FirmwareFlashDialog::createConnections()
 {
     connect(ui->toolBtn_firm, SIGNAL(clicked()), this, SLOT(onActnToolbtnClicked()));
     connect(ui->btn_firmFlash, SIGNAL(clicked()), this, SLOT(onActnFlashBtnClicked()));
+    connect(ui->comboBox_firm, SIGNAL(currentIndexChanged(int)), this, SLOT(onActnComboBoxIndexChanged(int)));
 }
 
 void FirmwareFlashDialog::processCallBack(void *argv, short *value)
@@ -61,6 +67,15 @@ void FirmwareFlashDialog::processCallBack(void *argv, short *value)
     QProgressBar *pBar = static_cast<QProgressBar *>(argv);
     pBar->setValue(*value);
     qApp->processEvents();
+}
+
+void FirmwareFlashDialog::closeEvent(QCloseEvent *event)
+{
+    QDir dir(m_desPath);
+    if (dir.exists()) {
+        deleteDir(m_desPath);
+    }
+    event->accept();
 }
 
 QStringList FirmwareFlashDialog::getFilesFromExt(const QString &fileExt, const QString &filePath, int num)
@@ -112,11 +127,40 @@ void FirmwareFlashDialog::onActnToolbtnClicked()
     if (m_decompressPath.compare("") == 0) {
         return;
     }
+    QDir dir(m_desPath);
+    if (dir.exists()) {
+        deleteDir(m_desPath);
+    }
+    dir.mkdir(m_desPath);
+    FolderCompressor *folderComp = new FolderCompressor(0);
+    bool ok = folderComp->decompressFolder(m_decompressPath, m_desPath);
+    delete folderComp;
+    if (!ok) {
+        QMessageBox::warning(this, tr("Warning"), tr("Decompressing fails!"));
+        deleteDir(m_desPath);
+        return;
+    }
     QFileInfo fileInfo;
     fileInfo.setFile(m_decompressPath);
     ui->lineEdit_firm->setText(fileInfo.fileName());
     m_filePath = m_decompressPath;
     ui->widget_firm->setVisible(true);
+
+    QFile infoFile(m_desPath + "/infoFile.ini");
+    if (!infoFile.open(QIODevice::ReadOnly | QIODevice::Text)) {
+        QMessageBox::warning(this, tr("Warning"), tr("File open fails!"));
+        deleteDir(m_desPath);
+        return;
+    }
+    QTextStream in(&infoFile);
+    QString line = in.readLine();
+    QStringList lineList = line.split(":");
+    m_dspVersion = lineList.last();
+
+    line = in.readLine();
+    lineList = line.split(":");
+    m_fpgVersion = lineList.last();
+    onActnComboBoxIndexChanged(ui->comboBox_firm->currentIndex());
 }
 
 void FirmwareFlashDialog::onActnFlashBtnClicked()
@@ -127,21 +171,44 @@ void FirmwareFlashDialog::onActnFlashBtnClicked()
     }
     ui->progressBar_firm->setValue(0);
     ui->progressBar_firm->setVisible(true);
-    ui->infoDisplay_firm->insertPlainText("Decompressing files!");
-    FolderCompressor *folderComp = new FolderCompressor(0);
-    bool ok = folderComp->decompressFolder(m_decompressPath, m_desPath);
-    delete folderComp;
+    ui->infoDisplay_firm->insertPlainText(tr("1.Checking version!\n"));
+
+    SevDevice *dev = m_devList.at(ui->comboBox_firm->currentIndex());
+    DeviceIdHelper *idHelper = new DeviceIdHelper(dev->socketCom(), 0);
+    bool ok;
+    quint32 pwrId = idHelper->readPwrId(ok);
     if (!ok) {
-        ui->infoDisplay_firm->insertPlainText("Decompress fails!");
-        deleteDir(m_desPath);
-        ui->progressBar_firm->setVisible(false);
+        ui->infoDisplay_firm->insertPlainText(tr("Reading powerboard Id fails!\n"));
+        delete idHelper;
         return;
+    }
+    QString pwrStr = QString::number(pwrId);
+
+    quint32 ctrId = idHelper->readCtrId(ok);
+    if (!ok) {
+        ui->infoDisplay_firm->insertPlainText(tr("Reading controlboard Id fails!\n"));
+        delete idHelper;
+        return;
+    }
+    QString ctrStr = QString::number(ctrId);
+    delete idHelper;
+
+    DBManager *dbManager = new DBManager(GTUtils::databasePath() + "Version/", "root", "");
+    QStringList verList;
+    verList<<ctrStr<<m_dspVersion<<m_fpgVersion<<pwrStr;
+    ok = dbManager->checkValid(verList);
+    delete dbManager;
+    if (!ok) {
+        int ret = QMessageBox::question(0, tr("Warning"), tr("Version do not match. Are you sure to continue?"), QMessageBox::Yes | QMessageBox::No, QMessageBox::No);
+        if (ret == 0) {
+            return;
+        }
     }
     if (ui->checkBox_firmHex->isChecked()) {
         ok = downloadHexFile();
     }
     if (!ok) {
-        ui->infoDisplay_firm->insertPlainText("Downloading hex file fails!");
+        ui->infoDisplay_firm->insertPlainText(tr("Downloading hex file fails!\n"));
         deleteDir(m_desPath);
         ui->progressBar_firm->setVisible(false);
         return;
@@ -150,7 +217,7 @@ void FirmwareFlashDialog::onActnFlashBtnClicked()
         ok = downloadRpdFile();
     }
     if (!ok) {
-        ui->infoDisplay_firm->insertPlainText("Downloading rpd file fails!");
+        ui->infoDisplay_firm->insertPlainText(tr("Downloading rpd file fails!\n"));
         deleteDir(m_desPath);
         ui->progressBar_firm->setVisible(false);
         return;
@@ -159,14 +226,34 @@ void FirmwareFlashDialog::onActnFlashBtnClicked()
         ok = downloadXmlFiles();
     }
     if (!ok) {
-        ui->infoDisplay_firm->insertPlainText("Downloading xml files fail!");
+        ui->infoDisplay_firm->insertPlainText(tr("Downloading xml files fail!\n"));
         deleteDir(m_desPath);
         ui->progressBar_firm->setVisible(false);
         return;
     }
-    ui->infoDisplay_firm->insertPlainText("Downloading succeeds!");
+    ui->infoDisplay_firm->insertPlainText(tr("Downloading succeeds!\n"));
     deleteDir(m_desPath);
     ui->progressBar_firm->setVisible(false);
+}
+
+void FirmwareFlashDialog::onActnComboBoxIndexChanged(int index)
+{
+    quint16 year = 0;
+    quint16 day = 0;
+    DeviceIdHelper *idHelper = new DeviceIdHelper(m_devList.at(index)->socketCom(), 0);
+    bool ok = idHelper->readFpgaDate(year, day);
+    delete idHelper;
+    if (!ok) {
+        qDebug()<<"read err";
+        return;
+    }
+    QString fpgDate = QString("%1%2").arg(QString::asprintf("%04X", year)).arg(QString::asprintf("%04X", day));
+    qDebug()<<"fpgdate"<<fpgDate;
+    if (fpgDate.compare(m_fpgVersion) == 0) {
+        ui->checkBox_firmRpd->setChecked(false);
+    } else {
+        ui->checkBox_firmRpd->setChecked(true);
+    }
 }
 
 bool FirmwareFlashDialog::deleteDir(const QString &path)
@@ -191,10 +278,11 @@ bool FirmwareFlashDialog::deleteDir(const QString &path)
 
 bool FirmwareFlashDialog::downloadHexFile()
 {
-    ui->infoDisplay_firm->insertPlainText("Downloading hex file");
+    ui->infoDisplay_firm->insertPlainText(tr("2.Downloading hex file!\n"));
     QStringList hexList = getFilesFromExt("hex", m_desPath, 1);
     QString hexPath = hexList.at(0);
-    SevDevice* dev = m_devList->at(ui->comboBox_firm->currentIndex());
+    SevDevice* dev = m_devList.at(ui->comboBox_firm->currentIndex());
+    bool co = dev->isConnecting();
     if (!dev->isConnecting()) {
         QMessageBox::information(0, tr("Warning"), tr("please open the com first !"));
         return false;
@@ -211,7 +299,8 @@ bool FirmwareFlashDialog::downloadHexFile()
     }
     delete tree;
     for (int i = 0; i < dspNum; i++) {
-        qint16 ret = dev->socketCom()->downLoadDSPFLASH(i, hexPath.toStdWString(), processCallBack, (void *)ui->progressBar_firm);
+        ui->infoDisplay_firm->insertPlainText(tr("Downloading DSP %1.\n").arg(i + 1));
+        qint16 ret = dev->socketCom()->downLoadDSPFLASH(i, hexPath.toStdWString(), processCallBack, (void *)ui->progressBar_firm);        
         if (ret != 0) {
             return false;
         }
@@ -221,12 +310,12 @@ bool FirmwareFlashDialog::downloadHexFile()
 
 bool FirmwareFlashDialog::downloadRpdFile()
 {
-    SevDevice* dev = m_devList->at(ui->comboBox_firm->currentIndex());
+    SevDevice* dev = m_devList.at(ui->comboBox_firm->currentIndex());
     if (!dev->isConnecting()) {
         QMessageBox::information(0, tr("Warning"), tr("please open the com first !"));
         return false;
     }
-    ui->infoDisplay_firm->insertPlainText("Downloading rpd file");
+    ui->infoDisplay_firm->insertPlainText(tr("3.Downloading rpd file!\n"));
     QStringList rpdList = getFilesFromExt("rpd", m_desPath, 1);
     QString rpdPath = rpdList.at(0);
     int fpgNum;
@@ -254,11 +343,12 @@ bool FirmwareFlashDialog::downloadRpdFile()
 
 bool FirmwareFlashDialog::downloadXmlFiles()
 {
-    SevDevice* dev = m_devList->at(ui->comboBox_firm->currentIndex());
+    SevDevice* dev = m_devList.at(ui->comboBox_firm->currentIndex());
     if (!dev->isConnecting()) {
         QMessageBox::information(0, tr("Warning"), tr("please open the com first !"));
         return false;
     }
+    ui->infoDisplay_firm->insertPlainText(tr("4.Downloading xml files!\n"));
     QStringList fileNameList;
     fileNameList.append(m_desPath + "/" + FILENAME_XML_FLASHPRM);
     fileNameList.append(m_desPath + "/" + FILENAME_XML_RAMPRM0);
